split installation of n98-magerun-modules in separate file don't download tools again if they're still present (local dev caching)

+++ setup.sh
@@ -31,69 +31,35 @@ git clone https://github.com/ffuenf/MageTestStand "${BUILDENV}"
 
 mkdir -p ${WORKSPACE}/build/logs
 mkdir -p ${BUILDENV}/tools
-curl -s -L https://raw.githubusercontent.com/colinmollenhour/modman/master/modman -o ${BUILDENV}/tools/modman
-chmod +x ${BUILDENV}/tools/modman
-curl -s -L http://files.magerun.net/n98-magerun-latest.phar -o ${BUILDENV}/tools/n98-magerun
-chmod +x ${BUILDENV}/tools/n98-magerun
-mkdir -p ~/.n98-magerun/modules/
-if [ ! -d ~/.n98-magerun/modules/yireo ] ; then
-  git clone https://github.com/yireo/magerun-addons ~/.n98-magerun/modules/yireo
+if [ ! -f ${BUILDENV}/tools/modman ] ; then
+  curl -s -L https://raw.githubusercontent.com/colinmollenhour/modman/master/modman -o ${BUILDENV}/tools/modman
+  chmod +x ${BUILDENV}/tools/modman
 fi
-if [ ! -d ~/.n98-magerun/modules/mpmd ] ; then
-  git clone https://github.com/aoepeople/mpmd ~/.n98-magerun/modules/mpmd
+if [ ! -f ${BUILDENV}/tools/n98-magerun ] ; then
+  curl -s -L http://files.magerun.net/n98-magerun-latest.phar -o ${BUILDENV}/tools/n98-magerun
+  chmod +x ${BUILDENV}/tools/n98-magerun
 fi
-if [ ! -d ~/.n98-magerun/modules/magerun-commands ] ; then
-  git clone https://github.com/degdigital/magerun-commands ~/.n98-magerun/modules/magerun-commands
+${BUILDENV}/n98-magerun-modules.sh
+if [ ! -f ${BUILDENV}/tools/composer ] ; then
+  curl -s -L https://getcomposer.org/composer.phar -o ${BUILDENV}/tools/composer
+  chmod +x ${BUILDENV}/tools/composer
 fi
-if [ ! -d ~/.n98-magerun/modules/magerun-addons ] ; then
-  git clone https://github.com/peterjaap/magerun-addons ~/.n98-magerun/modules/magerun-addons
+if [ ! -f ${BUILDENV}/tools/phploc ] ; then
+  curl -s -L https://phar.phpunit.de/phploc.phar -o ${BUILDENV}/tools/phploc
+  chmod +x ${BUILDENV}/tools/phploc
 fi
-if [ ! -d ~/.n98-magerun/modules/magerun-creatuity ] ; then
-  git clone https://github.com/creatuity/magerun-creatuity ~/.n98-magerun/modules/magerun-creatuity
+if [ ! -f ${BUILDENV}/tools/ocular ] ; then
+  curl -s -L https://scrutinizer-ci.com/ocular.phar -o ${BUILDENV}/tools/ocular
+  chmod +x ${BUILDENV}/tools/ocular
 fi
-if [ ! -d ~/.n98-magerun/modules/magerun-module-cache-benchmark ] ; then
-  git clone https://github.com/cmuench/magerun-module-cache-benchmark ~/.n98-magerun/modules/magerun-module-cache-benchmark
+if [ ! -f ${BUILDENV}/tools/assert.sh ] ; then
+  curl -s -L https://raw.github.com/lehmannro/assert.sh/master/assert.sh -o ${BUILDENV}/tools/assert.sh
+  chmod +x ${BUILDENV}/tools/assert.sh
 fi
-if [ ! -d ~/.n98-magerun/modules/cmuench-magerun-addons ] ; then
-  git clone https://github.com/cmuench/cmuench-magerun-addons ~/.n98-magerun/modules/cmuench-magerun-addons
+if [ ! -f ${BUILDENV}/tools/magedownload ] ; then
+  curl -s -L http://magedownload.steverobbins.com/download/latest/magedownload.phar -o ${BUILDENV}/tools/magedownload
+  chmod +x ${BUILDENV}/tools/magedownload
 fi
-if [ ! -d ~/.n98-magerun/modules/kalenjordan-magerun-addons ] ; then
-  git clone https://github.com/kalenjordan/magerun-addons ~/.n98-magerun/modules/kalenjordan-magerun-addons
-fi
-if [ ! -d ~/.n98-magerun/modules/Webgriffe_Golive ] ; then
-  git clone https://github.com/aleron75/Webgriffe_Golive ~/.n98-magerun/modules/Webgriffe_Golive
-fi
-if [ ! -d ~/.n98-magerun/modules/ffuenf-download-remote-media ] ; then
-  git clone https://github.com/ffuenf/download-remote-media ~/.n98-magerun/modules/ffuenf-download-remote-media
-fi
-if [ ! -d ~/.n98-magerun/modules/sxmlsv ] ; then
-  git clone https://github.com/KamilBalwierz/sxmlsv ~/.n98-magerun/modules/sxmlsv
-fi
-if [ ! -d ~/.n98-magerun/modules/magerun-dataprofile ] ; then
-  git clone https://github.com/marcoandreini/magerun-dataprofile ~/.n98-magerun/modules/magerun-dataprofile
-fi
-if [ ! -d ~/.n98-magerun/modules/Magerun-DBClean ] ; then
-  git clone https://github.com/steverobbins/Magerun-DBClean ~/.n98-magerun/modules/Magerun-DBClean
-fi
-if [ ! -d ~/.n98-magerun/modules/magerun-modman ] ; then
-  git clone https://github.com/fruitcakestudio/magerun-modman ~/.n98-magerun/modules/magerun-modman
-fi
-if [ ! -d ~/.n98-magerun/modules/EAVCleaner ] ; then
-  git clone https://github.com/magento-hackathon/EAVCleaner ~/.n98-magerun/modules/EAVCleaner
-fi
-if [ ! -d ~/.n98-magerun/modules/magescan ] ; then
-  git clone https://github.com/steverobbins/magescan ~/.n98-magerun/modules/magescan
-fi
-curl -s -L https://getcomposer.org/composer.phar -o ${BUILDENV}/tools/composer
-chmod +x ${BUILDENV}/tools/composer
-curl -s -L https://phar.phpunit.de/phploc.phar -o ${BUILDENV}/tools/phploc
-chmod +x ${BUILDENV}/tools/phploc
-curl -s -L https://scrutinizer-ci.com/ocular.phar -o ${BUILDENV}/tools/ocular
-chmod +x ${BUILDENV}/tools/ocular
-curl -s -L https://raw.github.com/lehmannro/assert.sh/master/assert.sh -o ${BUILDENV}/tools/assert.sh
-chmod +x ${BUILDENV}/tools/assert.sh
-curl -s -L http://magedownload.steverobbins.com/download/latest/magedownload.phar -o ${BUILDENV}/tools/magedownload
-chmod +x ${BUILDENV}/tools/magedownload
 
 cp ${BUILDENV}/.n98-magerun.yaml ~/.n98-magerun.yaml
